test(web): merge the two Models suites, drop duplicate Providers tests
The Models page had two parallel suites — pages/__tests__/Models.test.tsx and
pages/Models/__tests__/Models.test.tsx — sharing 27 identically named tests.
They now live as one lineage in the per-page dir, split by concern:

  Models.rendering.test.tsx   view mode, provider scope, loading, rendering,
                              count label, API error handling
  Models.mutations.test.tsx   model interactions, bulk delete, nav badge

For each of the 27 shared titles the version on the shared mockAllDefaults
handlers is kept; it carries every assertion the older copy had and, for the
models-API-error and delete-modal cases, more. The 12 tests only the older
file had (provider filter dropdown, provider scope, nav badge) and the 5 only
the newer one had (cursor total, enable toast, plural base key, bulk delete)
all survive: 71 tests become 44 with no assertion lost.

Providers had three tests duplicated across its two files. The copy in the
per-page file is kept and takes over the stronger assertions from the copy in
pages/__tests__: the exact empty-state sentence, and the full delete-failure
toast text. That toast reads "Failed to delete: Failed to delete provider"
because api.providers.delete discards the response body, so the comment now
says so instead of claiming the HTTP response is included.
