+++ web/src/pages/Models/__tests__/Models.mutations.test.tsx
@@ -2,8 +2,9 @@ import { screen, waitFor, within } from "@testing-library/react";
 import { HttpResponse, http } from "msw";
 import { beforeEach, describe, expect, it } from "vitest";
 import type { Model } from "../../../api/types";
-import { mockAllDefaults, mockModelsCursor } from "../../../test/helpers";
-import { mockModel } from "../../../test/mocks/data";
+import { Layout } from "../../../components/Layout";
+import { mockAllDefaults } from "../../../test/helpers";
+import { mockModel, mockProvider } from "../../../test/mocks/data";
 import { server } from "../../../test/mocks/server";
 import { renderWithProviders } from "../../../test/utils";
 import { Models } from "../../Models";
@@ -12,271 +13,6 @@ describe("Models", () => {
 	beforeEach(() => {
 		server.resetHandlers();
 		localStorage.setItem("modelsViewMode", "paginate");
-	});
-
-	describe("View Mode Toggle", () => {
-		it("starts in scroll mode by default and shows VirtualModelTable", async () => {
-			localStorage.removeItem("modelsViewMode");
-
-			server.use(...mockAllDefaults());
-
-			renderWithProviders(<Models />);
-
-			// Title should be "Models" (not count label)
-			await waitFor(() => {
-				expect(screen.getByText("Models")).toBeInTheDocument();
-			});
-
-			// Toggle glyph reflects scroll mode
-			expect(
-				screen
-					.getByTitle(
-						"Click to toggle between pagination and infinite scrolling.",
-					)
-					.querySelector(".icon-infinite-scroll"),
-			).toBeInTheDocument();
-
-			// Badge should not be shown in scroll mode
-			expect(screen.queryByText(/\d+ enabled/)).not.toBeInTheDocument();
-		});
-
-		it("switches from scroll to paginate mode when clicking toggle", async () => {
-			localStorage.removeItem("modelsViewMode");
-
-			server.use(...mockAllDefaults());
-
-			const { user } = renderWithProviders(<Models />);
-
-			await waitFor(() => {
-				expect(screen.getByText("Models")).toBeInTheDocument();
-			});
-
-			// Click toggle to switch to paginate mode
-			await user.click(
-				screen.getByTitle(
-					"Click to toggle between pagination and infinite scrolling.",
-				),
-			);
-
-			// Should now show count label
-			await waitFor(() => {
-				expect(screen.getByText("1 Model")).toBeInTheDocument();
-			});
-
-			// Toggle glyph now reflects paginate mode
-			expect(
-				screen
-					.getByTitle(
-						"Click to toggle between pagination and infinite scrolling.",
-					)
-					.querySelector(".icon-pages"),
-			).toBeInTheDocument();
-		});
-
-		it("switches from paginate to scroll mode when clicking toggle", async () => {
-			server.use(...mockAllDefaults());
-
-			const { user } = renderWithProviders(<Models />);
-
-			await waitFor(() => {
-				expect(screen.getByText("1 Model")).toBeInTheDocument();
-			});
-
-			// Click toggle to switch to scroll mode
-			await user.click(
-				screen.getByTitle(
-					"Click to toggle between pagination and infinite scrolling.",
-				),
-			);
-
-			// Should now show "Models" without count
-			await waitFor(() => {
-				expect(screen.getByText("Models")).toBeInTheDocument();
-			});
-
-			// Badge should not be shown in scroll mode
-			expect(screen.queryByText(/\d+ enabled/)).not.toBeInTheDocument();
-		});
-
-		it("does not show loading spinner in scroll mode even when models query is disabled", async () => {
-			localStorage.removeItem("modelsViewMode");
-
-			server.use(...mockAllDefaults());
-
-			renderWithProviders(<Models />);
-
-			// Should not show spinner - query is disabled in scroll mode
-			expect(screen.queryByTestId("spinner")).not.toBeInTheDocument();
-
-			// Should show title immediately
-			await waitFor(() => {
-				expect(screen.getByText("Models")).toBeInTheDocument();
-			});
-		});
-
-		it("shows model count in header when cursor API returns total", async () => {
-			localStorage.removeItem("modelsViewMode");
-
-			server.use(
-				...mockAllDefaults(),
-				...mockModelsCursor({
-					body: {
-						entries: [mockModel],
-						total: 42,
-						enabled_total: 42,
-						parked_total: 0,
-						disabled_total: 0,
-						has_before: false,
-						has_after: false,
-					},
-				}),
-			);
-
-			renderWithProviders(<Models />);
-
-			// Title should show count from cursor total
-			await waitFor(() => {
-				expect(screen.getByText("42 Models")).toBeInTheDocument();
-			});
-		});
-	});
-
-	describe("Loading State", () => {
-		it("renders loading spinner initially", () => {
-			server.use(
-				http.get("/api/models", () => {
-					return new Promise((resolve) => {
-						setTimeout(() => {
-							resolve(HttpResponse.json([mockModel]));
-						}, 100);
-					});
-				}),
-			);
-
-			renderWithProviders(<Models />);
-			expect(screen.getByTestId("spinner")).toBeInTheDocument();
-		});
-	});
-
-	describe("Rendering", () => {
-		it("renders page header with correct title and icon", async () => {
-			server.use(...mockAllDefaults());
-
-			renderWithProviders(<Models />);
-
-			await waitFor(() => {
-				expect(screen.getByText("1 Model")).toBeInTheDocument();
-			});
-			expect(
-				screen.getByText("Discovered models from your providers"),
-			).toBeInTheDocument();
-		});
-
-		it("renders model count badge with enabled/disabled breakdown", async () => {
-			const models = [
-				{ ...mockModel, id: "model-001", enabled: true },
-				{ ...mockModel, id: "model-002", enabled: true },
-				{ ...mockModel, id: "model-003", enabled: false },
-			];
-
-			server.use(...mockAllDefaults({ models }));
-
-			renderWithProviders(<Models />);
-
-			// Title counts usable rows (model AND provider enabled); the badge
-			// carries the remainder so the rows in view still add up.
-			await waitFor(() => {
-				expect(screen.getByText("2 Models")).toBeInTheDocument();
-			});
-			expect(screen.getByText("1 disabled")).toBeInTheDocument();
-			expect(screen.queryByText(/\d+ enabled/)).not.toBeInTheDocument();
-		});
-
-		it("renders model table with models", async () => {
-			server.use(...mockAllDefaults());
-
-			renderWithProviders(<Models />);
-
-			await waitFor(() => {
-				expect(screen.getByText("Test Model")).toBeInTheDocument();
-			});
-
-			// Table should have headers
-			expect(screen.getByText("Model")).toBeInTheDocument();
-			expect(screen.getByText("Capabilities")).toBeInTheDocument();
-			expect(screen.getByText("Provider")).toBeInTheDocument();
-			expect(screen.getByText("Discovered")).toBeInTheDocument();
-			expect(screen.getByText("Ctx")).toBeInTheDocument();
-			expect(screen.getByText("Max Out")).toBeInTheDocument();
-			expect(screen.getByText("Status")).toBeInTheDocument();
-		});
-
-		it("renders empty state when no models", async () => {
-			server.use(...mockAllDefaults({ models: [] }));
-
-			renderWithProviders(<Models />);
-
-			await waitFor(() => {
-				expect(
-					screen.getByText(
-						"No models discovered yet. Add a provider and discover models.",
-					),
-				).toBeInTheDocument();
-			});
-		});
-
-		it("renders model count in header correctly", async () => {
-			const models = Array.from({ length: 5 }, (_, i) => ({
-				...mockModel,
-				id: `model-${i}`,
-				model_id: `test-model-${i}`,
-			}));
-
-			server.use(...mockAllDefaults({ models }));
-
-			renderWithProviders(<Models />);
-
-			await waitFor(() => {
-				expect(screen.getByText("5 Models")).toBeInTheDocument();
-			});
-		});
-
-		it("shows all models enabled badge when all are enabled", async () => {
-			const models = [
-				{ ...mockModel, id: "model-001", enabled: true },
-				{ ...mockModel, id: "model-002", enabled: true },
-			];
-
-			server.use(...mockAllDefaults({ models }));
-
-			renderWithProviders(<Models />);
-
-			await waitFor(() => {
-				expect(screen.getByText("2 Models")).toBeInTheDocument();
-			});
-
-			// No breakdown badge when all same state
-			expect(screen.queryByText(/\d+ enabled/)).not.toBeInTheDocument();
-		});
-
-		it("shows all models disabled badge when all are disabled", async () => {
-			const models = [
-				{ ...mockModel, id: "model-001", enabled: false },
-				{ ...mockModel, id: "model-002", enabled: false },
-			];
-
-			server.use(...mockAllDefaults({ models }));
-
-			renderWithProviders(<Models />);
-
-			// Nothing usable: the title has no count and the badge explains why.
-			await waitFor(() => {
-				expect(screen.getByText("2 disabled")).toBeInTheDocument();
-			});
-			expect(
-				screen.getByRole("heading", { name: "Models" }),
-			).toBeInTheDocument();
-		});
 	});
 
 	describe("Model Interactions", () => {
@@ -857,94 +593,6 @@ describe("Models", () => {
 		});
 	});
 
-	describe("countLabel", () => {
-		it("shows 'Models' (without count) when 0 models in paginate mode", async () => {
-			// Ensure paginate mode is set
-			localStorage.setItem("modelsViewMode", "paginate");
-
-			server.use(...mockAllDefaults({ models: [] }));
-
-			renderWithProviders(<Models />);
-
-			// countLabel returns just "Models" for 0 count (not "0 Models")
-			await waitFor(() => {
-				expect(screen.getByText("Models")).toBeInTheDocument();
-			});
-
-			// Verify paginate mode is active (toggle shows the pages glyph)
-			expect(
-				screen
-					.getByTitle(
-						"Click to toggle between pagination and infinite scrolling.",
-					)
-					.querySelector(".icon-pages"),
-			).toBeInTheDocument();
-		});
-
-		it("passes the plural base key, so the header pluralises the noun", async () => {
-			// The header must hand countLabel "models.page_title" and let
-			// i18next choose the suffix. Handing it an already-resolved form
-			// (t("models.page_title_one")) still typechecks, but that string is
-			// not a key, so the title would read "2 Model" here and no locale
-			// with a third plural category could ever reach its own form.
-			localStorage.setItem("modelsViewMode", "paginate");
-
-			server.use(
-				...mockAllDefaults({
-					models: [
-						{ ...mockModel, id: "model-alpha", model_id: "alpha" },
-						{ ...mockModel, id: "model-beta", model_id: "beta" },
-					],
-				}),
-			);
-
-			renderWithProviders(<Models />);
-
-			await waitFor(() => {
-				expect(screen.getByText("2 Models")).toBeInTheDocument();
-			});
-		});
-	});
-
-	describe("API Error Handling", () => {
-		it("handles models API error gracefully", async () => {
-			server.use(
-				...mockAllDefaults({
-					models: { status: 500, body: { error: "Failed to fetch" } },
-				}),
-			);
-
-			renderWithProviders(<Models />);
-
-			// On query error, models is undefined, so models ?? [] = []
-			// Component renders empty state
-			await waitFor(() => {
-				expect(
-					screen.getByText(
-						"No models discovered yet. Add a provider and discover models.",
-					),
-				).toBeInTheDocument();
-			});
-		});
-
-		it("handles providers API error gracefully", async () => {
-			server.use(
-				...mockAllDefaults({
-					providers: { status: 500, body: { error: "Failed to fetch" } },
-				}),
-			);
-
-			renderWithProviders(<Models />);
-
-			await waitFor(() => {
-				expect(screen.getByText("Test Model")).toBeInTheDocument();
-			});
-
-			// Should still render models without provider data
-			expect(screen.getByText("Test Model")).toBeInTheDocument();
-		});
-	});
-
 	describe("handleDeleteDisabled", () => {
 		it("deletes all disabled models successfully and shows success toast", async () => {
 			const models = [
@@ -1045,6 +693,263 @@ describe("Models", () => {
 			await waitFor(() => {
 				expect(screen.getByText(/Failed to delete/)).toBeInTheDocument();
 			});
+		});
+	});
+
+	/**
+	 * The Models nav badge lives in Layout, on its own 60s poll on
+	 * ["discovery-status"]: this page's `["models"]` invalidations never reach
+	 * it. So these mount the page inside the real Layout and read the real badge
+	 * rather than a probe query, and each serves a `claim_count` that actually
+	 * falls once the write lands — a fixed payload would be right by accident
+	 * whether or not anything re-read it.
+	 */
+	describe("Models nav badge", () => {
+		const status = (claim_count: number) => ({
+			claims: [],
+			group_claims: [],
+			informational: [],
+			claim_count,
+			informational_unseen: 0,
+		});
+
+		const disabledModel = {
+			...mockModel,
+			id: "model-002",
+			model_id: "test-model-v2",
+			name: "Retired Model",
+			display_name: "Retired Model v2",
+			enabled: false,
+		};
+
+		function renderInLayout() {
+			return renderWithProviders(
+				<Layout>
+					<Models />
+				</Layout>,
+			);
+		}
+
+		/** Opens the detail modal for the enabled mock model. */
+		async function openDetailModal(
+			user: ReturnType<typeof renderWithProviders>["user"],
+		) {
+			await user.click(
+				(await screen.findByText("Test Model")).closest("tr") as HTMLElement,
+			);
+			return screen.findByRole("dialog");
+		}
+
+		it("re-reads the badge after a model is deleted", async () => {
+			// Deleting a gone model is one way to resolve its claim, and it is done
+			// from this page with the badge in view.
+			let deleted = false;
+			server.use(
+				http.get("/api/models", () =>
+					HttpResponse.json(deleted ? [] : [mockModel]),
+				),
+				http.get("/api/providers", () => HttpResponse.json([mockProvider])),
+				http.get("/api/discovery/status", () =>
+					HttpResponse.json(status(deleted ? 0 : 1)),
+				),
+				http.delete("/api/models/:id", () => {
+					deleted = true;
+					return new HttpResponse(null, { status: 204 });
+				}),
+			);
+
+			const { user } = renderInLayout();
+			expect(
+				await screen.findByTestId("discovery-status-badge"),
+			).toHaveTextContent("1");
+
+			const modal = await openDetailModal(user);
+			await user.click(within(modal).getByRole("button", { name: "Delete" }));
+			await user.click(
+				within(modal).getByRole("button", { name: "Confirm delete" }),
+			);
+			await waitFor(() => expect(deleted).toBe(true));
+
+			// Gone entirely, which the stale poll response would never produce.
+			await waitFor(() =>
+				expect(screen.queryByTestId("discovery-status-badge")).toBeNull(),
+			);
+		});
+
+		it("clears the table row too when a delete reports failure", async () => {
+			// Re-reading only the badge would leave the page contradicting itself:
+			// the count drops to reflect a model that is gone while its row sits
+			// there claiming it still exists. Both re-read on settle.
+			let deleted = false;
+			server.use(
+				http.get("/api/models", () =>
+					HttpResponse.json(deleted ? [] : [mockModel]),
+				),
+				http.get("/api/providers", () => HttpResponse.json([mockProvider])),
+				http.get("/api/discovery/status", () =>
+					HttpResponse.json(status(deleted ? 0 : 1)),
+				),
+				http.delete("/api/models/:id", () => {
+					deleted = true;
+					return HttpResponse.json({ error: "boom" }, { status: 500 });
+				}),
+			);
+
+			const { user } = renderInLayout();
+			expect(
+				await screen.findByTestId("discovery-status-badge"),
+			).toHaveTextContent("1");
+
+			const modal = await openDetailModal(user);
+			await user.click(within(modal).getByRole("button", { name: "Delete" }));
+			await user.click(
+				within(modal).getByRole("button", { name: "Confirm delete" }),
+			);
+			await screen.findByText(/Failed to delete/);
+
+			await waitFor(() => expect(screen.queryByText("Test Model")).toBeNull());
+			expect(screen.queryByTestId("discovery-status-badge")).toBeNull();
+		});
+
+		it("re-reads the badge after the disabled models are bulk deleted", async () => {
+			let deleted = false;
+			server.use(
+				http.get("/api/models", () =>
+					HttpResponse.json(deleted ? [mockModel] : [mockModel, disabledModel]),
+				),
+				http.get("/api/providers", () => HttpResponse.json([mockProvider])),
+				http.get("/api/discovery/status", () =>
+					HttpResponse.json(status(deleted ? 1 : 2)),
+				),
+				http.post("/api/models/bulk-delete", () => {
+					deleted = true;
+					return HttpResponse.json({ requested: 1, deleted: 1 });
+				}),
+			);
+
+			const { user } = renderInLayout();
+			expect(
+				await screen.findByTestId("discovery-status-badge"),
+			).toHaveTextContent("2");
+
+			await user.click(await screen.findByText("Delete 1 disabled"));
+			// ConfirmDialog confirms through the modal's fade-out, so the request is
+			// a timer away from the click, not a microtask.
+			await user.click(within(screen.getByRole("dialog")).getByText("Delete"));
+			await waitFor(() => expect(deleted).toBe(true));
+
+			// The enabled model's claim survives, so this pins a re-read rather than
+			// a badge that merely happens to be empty.
+			await waitFor(() =>
+				expect(screen.getByTestId("discovery-status-badge")).toHaveTextContent(
+					"1",
+				),
+			);
+		});
+
+		it("re-reads the badge after a model is toggled", async () => {
+			// Toggling reclassifies rather than removes: buildProviderClaims puts an
+			// enabled model in Suspect, and Suspect is not counted, so re-enabling a
+			// Gone model drops claim_count without deleting anything.
+			let enabled = false;
+			server.use(
+				http.get("/api/models", () =>
+					HttpResponse.json([{ ...mockModel, enabled }]),
+				),
+				http.get("/api/providers", () => HttpResponse.json([mockProvider])),
+				http.get("/api/discovery/status", () =>
+					HttpResponse.json(status(enabled ? 1 : 2)),
+				),
+				http.patch("/api/models/:id", () => {
+					enabled = true;
+					return HttpResponse.json({ ...mockModel, enabled: true });
+				}),
+			);
+
+			const { user } = renderInLayout();
+			expect(
+				await screen.findByTestId("discovery-status-badge"),
+			).toHaveTextContent("2");
+
+			const modal = await openDetailModal(user);
+			await user.click(
+				within(modal).getByRole("button", { name: /Enabled|Disabled/i }),
+			);
+			await waitFor(() => expect(enabled).toBe(true));
+
+			await waitFor(() =>
+				expect(screen.getByTestId("discovery-status-badge")).toHaveTextContent(
+					"1",
+				),
+			);
+		});
+
+		it("re-reads the badge after a bulk delete that reports failure", async () => {
+			// A rejected request does not prove the write did not land: the server
+			// can commit and the response be lost. The toast reports the failure,
+			// but the badge must not keep asserting a count it can no longer back.
+			let deleted = false;
+			server.use(
+				http.get("/api/models", () =>
+					HttpResponse.json(deleted ? [mockModel] : [mockModel, disabledModel]),
+				),
+				http.get("/api/providers", () => HttpResponse.json([mockProvider])),
+				http.get("/api/discovery/status", () =>
+					HttpResponse.json(status(deleted ? 1 : 2)),
+				),
+				http.post("/api/models/bulk-delete", () => {
+					deleted = true;
+					return HttpResponse.json({ error: "boom" }, { status: 500 });
+				}),
+			);
+
+			const { user } = renderInLayout();
+			expect(
+				await screen.findByTestId("discovery-status-badge"),
+			).toHaveTextContent("2");
+
+			await user.click(await screen.findByText("Delete 1 disabled"));
+			await user.click(within(screen.getByRole("dialog")).getByText("Delete"));
+			await screen.findByText(/Failed to delete/);
+
+			await waitFor(() =>
+				expect(screen.getByTestId("discovery-status-badge")).toHaveTextContent(
+					"1",
+				),
+			);
+		});
+
+		it("re-reads the badge after a discover run from the detail modal", async () => {
+			// A discover run can clear a claim by finding the model listed again, or
+			// raise one by confirming it missing. Either way the badge is as stale
+			// as it is after a dismissal.
+			let discovered = false;
+			server.use(
+				http.get("/api/models", () => HttpResponse.json([mockModel])),
+				http.get("/api/providers", () => HttpResponse.json([mockProvider])),
+				http.get("/api/discovery/status", () =>
+					HttpResponse.json(status(discovered ? 0 : 3)),
+				),
+				http.post("/api/providers/:id/discover", () => {
+					discovered = true;
+					return HttpResponse.json({ discovered: 1, diff: {} });
+				}),
+			);
+
+			const { user } = renderInLayout();
+			expect(
+				await screen.findByTestId("discovery-status-badge"),
+			).toHaveTextContent("3");
+
+			const modal = await openDetailModal(user);
+			await user.click(
+				within(modal).getByRole("button", { name: "Update info" }),
+			);
+			await waitFor(() => expect(discovered).toBe(true));
+
+			await waitFor(() =>
+				expect(screen.queryByTestId("discovery-status-badge")).toBeNull(),
+			);
 		});
 	});
 });
